api/requirements-simple.txt
fastapi
uvicorn
python-dotenv
groq
pandas
numpy
python-multipart
pydantic
aiofiles
requests
